bot.py
#!/usr/bin/env python3
"""
Repair Marketplace Telegram Bot
Production version with full functionality
"""

import asyncio
import logging
from typing import Optional
from datetime import datetime

from aiogram import Bot, Dispatcher, types
from aiogram.fsm.storage.redis import RedisStorage
from aiogram.webhook.aiohttp_server import SimpleRequestHandler, setup_application
from aiohttp import web

from config.settings import Config
from config.logging_config import setup_logging
from core.database import Database
from core.redis_client import RedisCache
from core.rabbitmq import RabbitMQClient
from handlers import (
    start, auth, search, booking, payment,
    profile, reviews, support, admin
)
from utils.middlewares import (
    AuthMiddleware, LoggingMiddleware, ThrottlingMiddleware
)
from utils.helpers import setup_webhook

import telebot
from telebot import types
import sqlite3
import schedule
import time
import threading
from datetime import datetime, timedelta
import google.auth
from google.oauth2.credentials import Credentials
from googleapiclient.discovery import build
import smtplib
from email.mime.text import MIMEText
from email.mime.multipart import MIMEMultipart
import requests
import json
import numpy as np
from sklearn.feature_extraction.text import TfidfVectorizer
from sklearn.naive_bayes import MultinomialNB
import pickle
import re

# Конфигурация
TOKEN = 'YOUR_TELEGRAM_BOT_TOKEN'
GOOGLE_CALENDAR_ID = 'YOUR_CALENDAR_ID'
OUTLOOK_CLIENT_ID = 'YOUR_OUTLOOK_CLIENT_ID'
OUTLOOK_CLIENT_SECRET = 'YOUR_OUTLOOK_CLIENT_SECRET'
SMTP_SERVER = 'smtp.gmail.com'
SMTP_PORT = 587
SMTP_USERNAME = 'your_email@gmail.com'
SMTP_PASSWORD = 'your_password'

bot = telebot.TeleBot(TOKEN)

# Настройка логирования
logger = setup_logging()

class RepairMarketplaceBot:
    """Главный класс бота"""
    
    def __init__(self):
        self.config = Config()
        self.bot = Bot(token=self.config.BOT_TOKEN)
        self.storage = RedisStorage.from_url(self.config.REDIS_URL)
        self.dp = Dispatcher(storage=self.storage)
        
        # Инициализация сервисов
        self.db = Database()
        self.cache = RedisCache()
        self.queue = RabbitMQClient()
        
        # Регистрация middleware
        self.setup_middlewares()
        
        # Регистрация обработчиков
        self.setup_handlers()
        
        logger.info("Bot initialized successfully")
    
    def setup_middlewares(self):
        """Настройка middleware"""
        self.dp.message.middleware(LoggingMiddleware())
        self.dp.message.middleware(ThrottlingMiddleware(rate_limit=0.5))
        self.dp.callback_query.middleware(AuthMiddleware())
    
    def setup_handlers(self):
        """Регистрация всех обработчиков"""
        
        # Базовые команды
        self.dp.include_router(start.router)
        self.dp.include_router(auth.router)
        
        # Основной функционал
        self.dp.include_router(search.router)
        self.dp.include_router(booking.router)
        self.dp.include_router(payment.router)
        self.dp.include_router(profile.router)
        self.dp.include_router(reviews.router)
        
        # Поддержка и администрирование
        self.dp.include_router(support.router)
        self.dp.include_router(admin.router)
    
    async def on_startup(self):
        """Действия при запуске"""
        logger.info("Starting bot...")
        
        # Подключение к сервисам
        await self.db.connect()
        await self.cache.connect()
        await self.queue.connect()
        
        # Настройка webhook
        if self.config.WEBHOOK_URL:
            await setup_webhook(
                self.bot,
                self.config.WEBHOOK_URL,
                self.config.WEBHOOK_PATH
            )
        
        # Запуск фоновых задач
        asyncio.create_task(self.run_background_tasks())
        
        logger.info("Bot started successfully")
    
    async def on_shutdown(self):
        """Действия при остановке"""
        logger.info("Shutting down bot...")
        
        # Закрытие соединений
        await self.db.disconnect()
        await self.cache.close()
        await self.queue.close()
        await self.bot.session.close()
        
        logger.info("Bot stopped")
    
    async def run_background_tasks(self):
        """Фоновые задачи"""
        while True:
            try:
                # Проверка просроченных заказов
                await self.check_expired_orders()
                
                # Отправка напоминаний
                await self.send_reminders()
                
                # Обновление кэша
                await self.update_cache()
                
                await asyncio.sleep(60)  # Каждую минуту
                
            except Exception as e:
                logger.error(f"Background task error: {e}")
                await asyncio.sleep(5)
    
    async def check_expired_orders(self):
        """Проверка просроченных заказов"""
        # Логика проверки
        pass
    
    async def send_reminders(self):
        """Отправка напоминаний"""
        # Логика отправки
        pass
    
    async def update_cache(self):
        """Обновление кэша"""
        # Логика обновления
        pass
    
    async def start_polling(self):
        """Запуск в режиме polling"""
        await self.on_startup()
        try:
            await self.dp.start_polling(self.bot)
        finally:
            await self.on_shutdown()
    
    async def start_webhook(self):
        """Запуск в режиме webhook"""
        app = web.Application()
        
        # Создание обработчика webhook
        webhook_requests_handler = SimpleRequestHandler(
            dispatcher=self.dp,
            bot=self.bot,
        )
        
        # Регистрация обработчика
        webhook_requests_handler.register(app, path=self.config.WEBHOOK_PATH)
        
        # Настройка приложения
        setup_application(app, self.dp, bot=self.bot)
        
        # Запуск сервера
        runner = web.AppRunner(app)
        await runner.setup()
        site = web.TCPSite(runner, '0.0.0.0', self.config.WEBHOOK_PORT)
        await site.start()
        
        await self.on_startup()
        
        # Ожидание завершения
        await asyncio.Event().wait()

def main():
    """Точка входа"""
    bot = RepairMarketplaceBot()
    
    # Выбор режима запуска
    if os.getenv('USE_WEBHOOK', 'false').lower() == 'true':
        asyncio.run(bot.start_webhook())
    else:
        asyncio.run(bot.start_polling())

if __name__ == '__main__':
    main()