requirements.txt.txt
# Core
aiogram==3.3.0
aiohttp==3.9.1
asyncio==3.4.3

# Database
aiomysql==0.2.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1

# Message Queue
aio-pika==9.3.0
celery==5.3.4

# AI/ML
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.4
transformers==4.36.0
torch==2.1.1
openai==1.6.1
langchain==0.0.340

# Payments
stripe==7.5.0
paypalrestsdk==1.13.1

# Google APIs
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Maps & Geo
geopy==2.4.0
folium==0.15.1
googlemaps==4.10.0

# Parsing
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.16.0
aiofiles==23.2.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0

# Utils
python-dotenv==1.0.0
pydantic==2.5.2
ujson==5.9.0
httpx==0.25.2
pillow==10.1.0
qrcode==7.4.2

aiogram==3.3.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
aiofiles==23.2.1
pydantic==2.5.0
# Для CRM, календарей добавьте позже:
# google-api-python-client
# requests
# scikit-learn (для ML)
# ...

# Telegram бот
aiogram==3.3.0

# База данных
SQLAlchemy==2.0.23
aiosqlite==0.19.0                # для SQLite (разработка)
asyncpg==0.29.0                   # для PostgreSQL (продакшен)

# Конфигурация
python-dotenv==1.0.0

# Работа с файлами
aiofiles==23.2.1

# Валидация данных
pydantic==2.5.0

# Дата и время
pytz==2023.3
tzlocal==5.2

# Логирование (опционально, удобнее стандартного)
loguru==0.7.2

# Google Maps
googlemaps==4.10.0

# Google Calendar
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Outlook Calendar (Microsoft Graph)
msal==1.24.0
requests==2.31.0                  # уже может быть установлена как зависимость

# ЮKassa
yookassa==2.3.2

# DonationAlerts (использует requests)
# дополнительных библиотек не требуется

# Машинное обучение
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.3

# Аналитика (план/факт, касса)
pandas==2.1.4

# Telegram клиент (Telethon)
Telethon==1.34.0

# Instagram (неофициальный API instagrapi)
instagrapi==2.0.0

# Планировщик задач (альтернатива Celery для простых проектов)
APScheduler==3.10.4

# Celery + Redis (для распределённых задач)
celery==5.3.4
redis==5.0.1

# Для REST API обычно достаточно requests
# При необходимости можно добавить специальные SDK:
# amoCRM API (неофициальный клиент)
amocrm-api-client==1.1.0

# Bitrix24 REST API клиент
bitrix24-python-sdk==1.0.0

pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# AI/ML
scikit-learn==1.3.2
pandas==2.1.4
joblib==1.3.2
numpy==1.24.3

# Парсинг и поиск
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3

# Telethon (для поиска групп)
Telethon==1.34.0
cryptg==0.4.0  # для ускорения

# instagrapi (осторожно, неофициальное API)
instagrapi==2.0.0

# Для SEO (анализ конкурентов)
# (дополнительных библиотек не требуется)

# Core
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.7.0
python-dotenv==1.0.0
requests==2.31.0
httpx==0.27.0

# Database
asyncmy==0.2.9
sqlalchemy==2.0.30
alembic==1.13.1
redis==5.0.4
aioredis==2.0.1

# AI/ML
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
transformers==4.40.0
torch==2.3.0
openai==1.30.0
langchain==0.2.3

# Payments
stripe==9.0.0
paypalrestsdk==1.13.1

# Google APIs
google-api-python-client==2.126.0
google-auth-oauthlib==1.2.0
googlemaps==4.10.0

# Utils
celery==5.3.6
aio-pika==9.4.1
python-multipart==0.0.9
pyjwt==2.8.0
cryptography==42.0.7